flask_6_api_management/myapp/azure_deployment/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
jsonify
azure-functions